Server/requirements.txt
asgiref==3.7.2
boto3==1.28.72
botocore==1.31.72
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Deprecated==1.2.14
distlib==0.3.7
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.57.0
django-braces==1.15.0
django-cors-headers==4.3.0
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.14.2
djangorestframework==3.14.0
environ==1.0
filelock==3.12.3
future==0.18.3
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
jwcrypto==1.5.0
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.1.0
pipenv==2023.9.1
platformdirs==3.10.0
psycopg2==2.9.9
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.7.0
serializers==0.2.4
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==6.7.0
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.4
wrapt==1.15.0
